+++ Block/Checkout/Agreements.php
@@ -1,14 +1,4 @@
 <?php
-/*
- *
- *
- * File: Agreements.php
- * Modified on: 29/01/2021 16:41
- *
- * @author Vincent Pietri <vincent.pietri@agence-tbd.com>
- * @copyright 2014 - 2021 TBD, SAS. All rights reserved.
- */
-
 namespace Monext\Payline\Block\Checkout;